Adjust socket functions
